requirements.txt
pyserial
srp
digi-xbee
numpy<2.0.0
pandas
opencv-python
scikit-image
pyjpegls